Move "Read Rest of Output" Out of Destructor
With the "read rest of output" code in the Comm destructor, it would
continue to read output even in situations where some error happened and
we expect sploit to die or when the user presses Ctrl+C to end sploit.
By moving it to the end of the script running code in main, it behaves
more intuitively.

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>