Handle Exceptions in Daemon Mode
If the user presses Ctrl+C while waiting on a connection, we want to
gracefully exit.
If the user presses Ctrl+C during the script, we want to stop executing
the script and restart the loop.
If any other exception happens during the script, we want to print out
the stacktrace as normal, but continue the loop.

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>