sploit: Properly scope user-script execution
Rather than implicitly inheriting names in scope for the user-script,
this collection is sanitized and we only export the 'comm' communication
object.  This seems to be a safer way to operate and addresses an issue
with sub-scopes in the user's script not functioning properly.
(Previously, user-defined functions did not have access to globals, or
library functions.)

Additionally, the user's code is now passed through compile() to attach
the original file name.  This is useful for debugging / diagnostic
situations, to make it more obvious if a crash originated from the
user's script.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>