sploit: Add function popen()
This is a free-function in the comm module, intended to help setup
Sploit plumbing when working in the Python interactive interpreter.

At the moment, the intended user experience in the interpreter is to err
on the side of being interactive/responsive.  As such, the Comm object
returned from popen() is initialized with overridden IO settings to
prefer 'readonwrite' by default.  Addtionally, any early output from the
target is also read, so that it may be immediately visible.

A consequence of this configuration is that, until readonwrite is set
False, most target output will be consumed before any .read* function
has a chance to return it.  While that would be a hard showstopper for
any Sploit script, an interactive user can simply copy/paste any
important data that is produced.  Given that the interpreter workflow is
likely going to be most useful for quick prototyping and recon with the
proposed rev module, I consider this tradeoff appropriate at the moment,
but will consider revisiting this if its usage is problematic.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>