sploit: reverse direction of r2 get_locals offsets
rev.r2's get_locals() function returns a Symtbl of offsets representing
the local variables on in a stack frame of a particular function.  The
offsets returned by r2 are based around the base of the stack, but they
are increasing in value as they grow from the stack.  To properly model
memory, they should decrease in value as they grow from the stack.

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>