sploit: Fix bugs and simplify Symtbl
The recent implementation of the new design for Symtbl contained a few
bugs:

        - Attempting to access .base on a Symtbl or intermediate
          __InnerTable__ caused an exception.

        - Symtbl objects all used the same static collection of nested
          subtables, rather than an instanced one.  If two table objects
          contained the same named key, they would refer to the same
          nested table from both locations.

        - Printing the contents of a table accessed via an absolute
          nesting (aka: via an __InnerTable__ object) would not show the
          offsets adjusted for the curent context.

In addition to these fixes, the class implementation is largely
simplified as well.  This is in part due to the removal of unnecessary
logic, such as the way our __getattribute__ overloads were implemented.
Mainly, this came down to merging the redundant abstractions in our
original design.

Over time, the differences between these interfaces became blurred to
the point where simply reusing one is not at all problematic.  It is
very much the intent of this patch to preserve the semantics of the
tool's design (that being: flexable, nestable tables, to which a
separate, but linked, mapped view may be obtained), but to state it as
cleanly as possible.

Note that all of the working state of a Symtbl is kept in its new
_namesp member.  This is primarily done to enable subclassing the Symtbl
class.  Ordinarily, setattr() on self would force the incoming value
into the actual symbol table, making it impossible for subclasses to
store separate instance data.  Furthermore, the consolidation of
properties into this object creates fewer potential collisions with
user-defined symbols.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>