sploit: mem: Allow Symtbl base to be modified
Allow a Symtbl's base to be modified in-place, without mapping into a
new object.  This is useful when working with the Symtbl aspect of a
Payload.

This includes setting a non-zero base on construction.  As usual, when
defining base on construction, any additional kwargs symbols are
interpreted relative to the given base.  The order of arguments does not
matter.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>