sploit: payload: Class no longer extends Symtbl
Given the current design of Symtbl, creating subclasses of it gets more
tedious the further one goes down a potential class hierarchy.  As I am
planning to introduce new features in the future that explicitly extend
Payload, make this change now to minimize the impact.

Additionally, switching Payload's relationship with Symtbl from "is-a"
to "has-a" makes it more consistent with rev.ELF, the other major user
of Symtbl.  (And in both cases, the member is named 'sym')

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>