comm: Strip \n character from readline()
Line-oriented reads now strip the newline from the end of their returned
string.  Additionally, readall() strips the newline, but only from the
string that gets logged to the user's terminal (goodbye to all the "\n"
printed at the end of each line).

Of course, these functions are called by other parts of the read API and
have downstream effects.  Consideration was given to the entire API with
these rules in mind:

    - Raw reads (or non-line-oriented reads) will not filter ANY of
      their read content.  They are logged to the screen as one "line"
      of log text with \n characters shown in-place (not actually
      resetting the terminal cursor).  If reading binary, these bytes
      dont actually mean line termination anyway.

      functions: read, readall(_nonblock) *, readuntil

    - Line-oriented reads will strip the terminating \n, log the single
      line to the screen, and return it.

      functions: readline, readlineuntil **

* readall(_nonblock) functions turn out to be a special case.  They will
operate as raw reads, returning a blob of content.  However, we
generally want to run them on line-oriented input, so they log according
to the line-oriented rules.

** Although content returned from readlineuntil will have \n's stripped,
the lines are returned in an array, so we can still distinguish them.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>