Use buffered read throughout Comm
We had originally decided to use the os.read() function instead of the
actual buffered file object's read function.  This was due to the
blocking behavior or os.read() being closer to POSIX read than the other
function.

As it turns out, os.read() is an unbuffered read.  Every other read call
in this interface is buffered.  This causes some undefined behavior in
certain cases and leads to some really confusing bugs.

After some discussion, we've decided that, in this application's domain,
the blocking behavior of the buffered file object's read is actually
often more useful anyways.  Changing this call will deal with both
issues.

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>