comm: Localize stdin nonblock to interact's readall
In interact(), we set stdin to be nonblocking for the duration of the
function.  As an unexpected side-effect, this was setting stdout to be
nonblocking as well.  This has caused at least one crash in the past.

Localizing the nonblock to just when we're reading from stdin should
solve this.

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>