symtbl: Rename file to match class name
I assume that the preferred style is to leave one major class each to a
file.  In this case, synchronize the names of the Symtbl class and its
containing module.  Per PEP8, the module is lowercase, and the class
remains Pascal case.

If other memory-oriented utilities are introduced in the future, we may
wish to move them, as well as Symtbl, back into a subpackage named
'mem'.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>