symtbl: Refactor module as an improved container type (and more)
This effort was triggered by three immediate wants of the module:

    An improved data container interface to support things like key
    iteration and better key management.  This is primarily wanted by
    the ROP module (which is still in development).

    The introduction of package documentation across the project.  This
    module is now fully documented.

    To fix a bug in the Symtbl constructor, which would not allow a
    caller to supply "self" as an initial symbol name, even though it is
    legal in every other context.  This problem was caused by the
    constructor's bound instance parameter sharing this name.

This patch addresses all of these concerns, and also introduces some
fringe / QoL improvements that were discovered during the API refactor.

Element access may now be done via subscripting, as well as the previous
(and still generally perferred) .attribute notation.  The syntax for
storing subtables within a parent Symtbl is now greatly streamlined due
to some implementation-level changes to the class.  You may now directly
assign just a Symtbl object or a normal int, and you don't have to fuss
with tuples anymore.  The subtable's base is taken as its offset in the
parent, and the new operator replacement for the .map() method may be
used to define a desired value for the parent.

This detail is actually a breaking change compared to the previous
version.  While not technically a bug, it is unintuitive that the
previous version would not remove subtables when their offset was
changed by a simple assignment - the table would just move.  This patch
make it such that any symbol assignment to a regular int will replace an
old mounted subtable if one exists.

There are now no normal instance methods on the Symtbl type (only dunder
method overrides).  This is to free up the available symbol namespace as
much as possible.  The previous methods map(), adjust(), and rebase()
are now implemented as operators which, in every case, yield a new
derivative object, rather than mutating the original.  All operators are
listed here:

    @   remap to absolute address
    +   remap to relative address
    -   remap to negated relative address
    >>  adjust all symbol offsets upward
    <<  adjust all symbol offsets downward
    %   rebase all symbol offsets around an absolute zero point

Additionally, Symtbl objects will convert to an integer via int(),
hex(), oct(), or bin(), yielding the base value.

The addition of these operators presents another breaking change to the
previous version.  Previously, symbol adjustments or rebases affected
the tracked offsets and caused symbols to shift around in linked tables
as well.  Since these operators now preserve the state of their source
object, this is no longer the case.  The amount of shift due to
adjustment or rebasing is localized in a specific Symtbl instance (and
is affected the the use of the related operators), however this value is
inherited by derivatives of that object.

There is a third breaking change caused by the use of operators as well.
Previously, the map() function allowed the caller to specify that the
given absolute address is not that of the table base, but of some offset
in the table, from which the new base is calculated.  However, the
remapping operators take only a single numeric value as their right hand
side operand, which is the absolute or relative address.  The new
intended way of accomplishing this (which is _nearly_ equivalent) is
through the combined use of the rebase and remap operations:

    # The address of the puts() function in a libc tbl is leaked
    sym = sym % sym.puts @ leak

aka: adjust offsets such that the known point is at the base, then move
that base to the known location.  The way in which this is different to
what you would end up with before is that previously, following a
map(abs, off) the base of the table would be accurately valued
according to the known information.  Now, the 'base' is considered to be
the leaked value, but internal offsets are shifted such that they still
resolve correctly.

Finally, a few new pieces of functionality are added to build out the
container API:

    - symbol key deletion
    - iteration over symbol:offset pairs
    - can now check for symbol existence with the "in" keyword
    - len(symtbl) returns the number of symbols defined

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>