Add special cases for read(size <= 0)
A read of 0 isn't particularly useful, but it is weird that it will
cause a BrokenPipeError. Instead, it makes more sense to just return an
empty string.

A read of <0 would normally read until EOF, but we already have that
feature in readall() and it wouldn't be particularly useful here.  A
similar functionality of reading the entire current contents of the
buffer is useful, though.  This is already implemented in
readall_nonblock() and this would be a nice user-facing way of calling
that.

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>