Prefer __repr__ for pretty-printing objects
Define human-readable string formatting for objects in repr, rather than
str, as this will enable an interactive interpreter to more conveniently
show this data to the user.  I believe this especially makes sense in
cases where __str__ doesn't perform a semantic type conversion for its
class (currently, all affected cases).

Scripts can still easily yield this information by using
`print(object)`, as print will fallback to repr(object) when there is
not an explicitly defined __str__.

Furthermore, this patch still maintains backwards compatability (for the
time being) of using str(object) to retrieve the information.  This is
because the default __str__ implementation will defer to __repr__ if
provided.  This made the Symtbl case of providing both of them
especially redundant.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>