symtbl: Overload __getitem__ for translating raw offsets
Can now use Symtbl subscript syntax to obtain the mapped address of a
foreign offset (not a defined symbol) without having to modify the
object or add a new symbol entry.

Assuming a base value of 10, tbl[15] will return 25, for example.

We now assert that the defined table keys are strings, to prevent the
creation of entries that are now un-readable by this patch.  However,
this always should have been the case.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>