Allow control of named pipe location via command-line
Add the ability to select which location to create FIFOs when running in
pipes mode, by passing the directory name to sploit where a target
executable would usually go.  This has been an API feature from the start,
but not exposed via the sploit runner command-line interface.

There are a couple new use-cases where this is very convenient, including
scriptifying sploit in pipes mode (testing, for example) and when running
sploit under Docker.  If pipes are placed in the working directory, all
project files can be shared with a single bind mount.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>